+++ spring-boot-email-core/src/main/java/it/ozimov/springboot/mail/service/TemplateService.java
@@ -17,7 +17,6 @@
 package it.ozimov.springboot.mail.service;
 
 import it.ozimov.springboot.mail.service.exception.TemplateException;
-import lombok.NonNull;
 
 import java.io.IOException;
 import java.util.Map;