+++ spring-boot-email-core/src/test/java/it/ozimov/springboot/mail/service/defaultimpl/EmailSchedulerPropertiesContextBasedTest.java
@@ -76,8 +76,6 @@ public class EmailSchedulerPropertiesContextBasedTest implements ContextBasedTes
                             SPRING_MAIL_SCHEDULER_PRIORITY_LEVELS + "=123",
                             SPRING_MAIL_PERSISTENCE_ENABLED + "=false",
                             SPRING_MAIL_PERSISTENCE_ENABLED + "=false",
-                            SPRING_MAIL_SCHEDULER_PERSISTENCE_REDIS_EMBEDDED + "=false",
-                            SPRING_MAIL_SCHEDULER_PERSISTENCE_REDIS_ENABLED + "=false",
                             SPRING_MAIL_SCHEDULER_PERSISTENCE_DESIRED_BATCH_SIZE + "=1",
                             SPRING_MAIL_SCHEDULER_PERSISTENCE_MIN_KEPT_IN_MEMORY + "=1",
                             SPRING_MAIL_SCHEDULER_PERSISTENCE_MAX_KEPT_IN_MEMORY + "=1"
@@ -102,45 +100,6 @@ public class EmailSchedulerPropertiesContextBasedTest implements ContextBasedTes
                     .hasFieldOrPropertyWithValue("priorityLevels", 123)
                     .hasFieldOrPropertyWithValue("persistence", EmailSchedulerProperties.Persistence.builder()
                             .desiredBatchSize(1).minKeptInMemory(1).maxKeptInMemory(1).build());
-        }
-
-    }
-
-
-    @RunWith(SpringRunner.class)
-    @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-    @TestPropertySource(locations = "classpath:application.properties",
-            properties =
-                    {
-                            SPRING_MAIL_SCHEDULER_ENABLED + "=true",
-                            SPRING_MAIL_SCHEDULER_PRIORITY_LEVELS + "=321",
-                            SPRING_MAIL_PERSISTENCE_ENABLED + "=true",
-                            SPRING_MAIL_SCHEDULER_PERSISTENCE_REDIS_EMBEDDED + "=false",
-                            SPRING_MAIL_SCHEDULER_PERSISTENCE_REDIS_ENABLED + "=false",
-                            SPRING_MAIL_SCHEDULER_PERSISTENCE_DESIRED_BATCH_SIZE + "=125",
-                            SPRING_MAIL_SCHEDULER_PERSISTENCE_MIN_KEPT_IN_MEMORY + "=25",
-                            SPRING_MAIL_SCHEDULER_PERSISTENCE_MAX_KEPT_IN_MEMORY + "=123456"
-                    })
-    public static class EmailSchedulerPropertiesWithPersistenceContextBasedTest implements ContextBasedTest {
-
-        @Rule
-        public Timeout timeout = new Timeout(10, TimeUnit.SECONDS);
-
-        @Rule
-        public final JUnitSoftAssertions assertions = new JUnitSoftAssertions();
-
-        @Autowired
-        private EmailSchedulerProperties emailSchedulerProperties;
-
-        @Test
-        public void shouldSchedulerPropertiesHaveExpectedValues() throws Exception {
-            //Assert
-            assertions.assertThat(emailSchedulerProperties)
-                    .isNotNull()
-                    .hasFieldOrPropertyWithValue("enabled", true)
-                    .hasFieldOrPropertyWithValue("priorityLevels", 321)
-                    .hasFieldOrPropertyWithValue("persistence", EmailSchedulerProperties.Persistence.builder()
-                            .desiredBatchSize(125).minKeptInMemory(25).maxKeptInMemory(123456).build());
         }
 
     }